# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Sistema de Gestao para as Oficinas Gengis Khan API
  version: 0.1.0
  description: API machine-readable baseline aligned with backend implementation.
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/v1/customers:
    get:
      tags: [customers]
      summary: List customers
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Customers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
    post:
      tags: [customers]
      summary: Create customer
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '200':
          description: Created customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
  /api/v1/customers/{nif}:
    parameters:
      - $ref: '#/components/parameters/NifParam'
    get:
      tags: [customers]
      summary: Get customer by NIF
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Customer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
    patch:
      tags: [customers]
      summary: Update customer
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated customer
  /api/v1/customers/{nif}/history:
    parameters:
      - $ref: '#/components/parameters/NifParam'
    get:
      tags: [customers]
      summary: Get customer history
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Customer history
  /api/v1/scooters:
    get:
      tags: [scooters]
      summary: List scooters
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Scooters list
    post:
      tags: [scooters]
      summary: Create scooter
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Scooter'
      responses:
        '200':
          description: Created scooter
  /api/v1/scooters/{serialNumber}:
    parameters:
      - $ref: '#/components/parameters/SerialNumberParam'
    get:
      tags: [scooters]
      summary: Get scooter
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Scooter detail
    patch:
      tags: [scooters]
      summary: Update scooter
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Updated scooter
...
  /api/v1/parts:
    get:
      tags: [inventory]
      summary: List parts
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Parts list
    post:
      tags: [inventory]
      summary: Create part
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Part created
  /api/v1/parts/low-stock:
    get:
      tags: [inventory]
      summary: List low stock parts
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Low stock parts
  /api/v1/parts/{partReference}:
    parameters:
      - name: partReference
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [inventory]
      summary: Get part
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Part detail
    patch:
      tags: [inventory]
      summary: Update part
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Updated part
  /api/v1/parts/{partReference}/stock-adjustments:
    parameters:
      - name: partReference
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [inventory]
      summary: Apply manual stock adjustment
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Part stock adjusted
  /api/v1/stock-movements:
    get:
      tags: [inventory]
      summary: List stock movements
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Stock movement history
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  parameters:
...
    NifParam:
      name: nif
      in: path
      required: true
      schema:
        type: string
    SerialNumberParam:
      name: serialNumber
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    Customer:
      type: object
      properties:
        nif:
          type: string
        customerType:
          type: string
          enum: [personal, business]
        fullName:
          type: string
        legalName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        address:
          type: string
        creditLimit:
          type: string
        paymentTerms:
          type: string
        isArchived:
          type: boolean
    Scooter:
      type: object
      properties:
        serialNumber:
          type: string
        brand:
          type: string
        model:
          type: string
        conditionNotes:
          type: string
        customerNif:
          type: string
        isArchived:
          type: boolean
...